requirements.txt
# ============================================================================
# MOMENT Foundation Model - Microservice Anomaly Detection
# Requirements for RTX 4070 (8GB VRAM) - Memory Optimized
# Python 3.8+
# ============================================================================

# Core Deep Learning Framework
torch>=2.0.0,<3.0.0              # PyTorch with CUDA support
torchvision>=0.15.0              # Vision utilities (if needed)
torchaudio>=2.0.0                # Audio utilities (minimal)

# MOMENT Foundation Model
momentfm>=0.1.0                  # Time-series foundation model (ICML 2024)

# HuggingFace Ecosystem
transformers>=4.35.0             # Transformer models and utilities
datasets>=2.14.0                 # Dataset loading and processing
accelerate>=0.24.0               # Distributed training and memory optimization
peft>=0.6.0                      # Parameter-Efficient Fine-Tuning (LoRA)
bitsandbytes>=0.41.0             # 8-bit quantization for memory savings

# Time Series Libraries
tsai>=0.3.5                      # Time series AI utilities
tsfresh>=0.20.0                  # Automated time series feature extraction
statsmodels>=0.14.0              # Statistical models (ARIMA, etc.)

# Classical ML (Baselines)
scikit-learn>=1.3.0              # Isolation Forest, Random Forest, metrics
xgboost>=2.0.0                   # Gradient Boosting (if needed)
catboost>=1.2.0                  # CatBoost (Phase 2 baseline)

# Data Processing
numpy>=1.24.0,<2.0.0             # Numerical computing
pandas>=2.0.0                    # Data manipulation
scipy>=1.10.0                    # Scientific computing

# Visualization
matplotlib>=3.7.0                # Basic plotting
seaborn>=0.12.0                  # Statistical visualization
plotly>=5.17.0                   # Interactive plots
tensorboard>=2.14.0              # Training visualization

# Evaluation and Metrics
scikit-metrics>=0.1.0            # Extended metrics (if available)
imbalanced-learn>=0.11.0         # Handling imbalanced data

# Configuration and Utilities
pyyaml>=6.0                      # YAML config files
python-dotenv>=1.0.0             # Environment variables
tqdm>=4.66.0                     # Progress bars
click>=8.1.0                     # CLI interface
loguru>=0.7.0                    # Advanced logging

# Memory Profiling (Development)
memory-profiler>=0.61.0          # Memory usage tracking
psutil>=5.9.0                    # System monitoring
gputil>=1.4.0                    # GPU monitoring

# Jupyter (Optional - for exploration)
jupyter>=1.0.0                   # Jupyter notebook
ipywidgets>=8.1.0                # Interactive widgets
jupyterlab>=4.0.0                # JupyterLab interface

# Code Quality (Optional - for development)
black>=23.10.0                   # Code formatting
isort>=5.12.0                    # Import sorting
flake8>=6.1.0                    # Linting
mypy>=1.6.0                      # Type checking
pytest>=7.4.0                    # Testing framework
pytest-cov>=4.1.0                # Coverage reporting

# ============================================================================
# Installation Instructions
# ============================================================================
#
# 1. Create virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    # OR
#    venv\Scripts\activate     # Windows
#
# 2. Install PyTorch with CUDA (check CUDA version first with: nvidia-smi):
#    # For CUDA 11.8:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#    # For CUDA 12.1:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#
# 3. Install remaining requirements:
#    pip install -r requirements.txt
#
# 4. Verify installation:
#    python -c "import torch; print(f'PyTorch: {torch.__version__}'); print(f'CUDA Available: {torch.cuda.is_available()}'); print(f'CUDA Version: {torch.version.cuda}')"
#    python -c "import momentfm; print(f'MOMENT: {momentfm.__version__}')"
#
# ============================================================================
# Memory Optimization Notes for RTX 4070 (8GB VRAM)
# ============================================================================
#
# - Use gradient_checkpointing=True in model config
# - Use mixed precision (torch.cuda.amp.autocast)
# - Start with batch_size=8, increase to 16 if stable
# - Use LoRA rank=8 (not 16) to reduce trainable parameters
# - Consider 8-bit quantization with bitsandbytes if OOM occurs
# - Monitor GPU memory with: watch -n 0.5 nvidia-smi
#
# ============================================================================
